Check if label files exist v2 (#25)
* Check if label files exist.

* Add vscode cache file to .gitignore

Fixes: https://github.com/spinalcordtoolbox/manual-correction/issues/23

+++ package_for_correction.py
@@ -147,9 +147,6 @@ def main():
     # Curate dict_yml to only have filenames instead of absolute path
     dict_yml = utils.curate_dict_yml(dict_yml)
 
-    # Check for missing files before starting the whole process
-    utils.check_files_exist(dict_yml, utils.get_full_path(args.path_in))
-
     suffix_dict = {
         'FILES_SEG': args.suffix_files_seg,         # e.g., _seg or _label-SC_mask
         'FILES_GMSEG': args.suffix_files_gmseg,     # e.g., _gmseg or _label-GM_mask
@@ -157,6 +154,9 @@ def main():
         'FILES_LABEL': args.suffix_files_label,     # e.g., _labels or _labels-disc
         'FILES_PMJ': args.suffix_files_pmj          # e.g., _pmj or _label-pmj
     }
+
+    # Check for missing files before starting the whole process
+    utils.check_files_exist(dict_yml, utils.get_full_path(args.path_in), suffix_dict)
 
     # Create temp folder
     path_tmp = tempfile.mkdtemp()
@@ -186,15 +186,18 @@ def main():
             # Construct absolute path to the temp folder
             path_out = os.path.join(path_tmp, subject, ses, contrast)
             # Copy image
-            copy_file(fname, path_out)
+            if os.path.exists(fname):
+                copy_file(fname, path_out)
             if args.other_contrast:
-                copy_file(fname_other_contrast, path_out)
+                if os.path.exists(fname_other_contrast):
+                    copy_file(fname_other_contrast, path_out)
             # Copy label if exists
             if suffix_label is not None:
                 # Construct absolute path to the input label (segmentation, labeling etc.) file
                 # For example: '/Users/user/dataset/data_processed/sub-001/anat/sub-001_T2w_seg.nii.gz'
                 fname_seg = utils.add_suffix(fname, suffix_dict[task])
-                copy_file(fname_seg, path_out)
+                if os.path.exists(fname_seg):
+                    copy_file(fname_seg, path_out)
 
     # Package to zip file
     print("Creating archive...")
